This slightly over-large commit implements port deletion & renaming Resolves issues: 3069, 3070, 2862, 3381
Major Changes:
 + Renamed several exposed fn's to be Splice-consistent, (DeletePort -> RemovePort, NumPorts -> PortCount
 + Updated editor to match latest namings
 + Added Splice.LoadExtension MxS function, which can be used to reload extensions
 + Fixed several crashes caused by no input checking in editor
 + Impelemented parameter renaming
 + Fixed changing max parameter

+++ src/SpliceEvents.cpp
@@ -171,7 +171,7 @@ public:
 				float fovX = pView->GetFOV();
 				//convert to vertical fov as the RTR camera is always using this mode:
 				double aspect = double(width) / double(height);
-				float fovY = (2.0 * atan(1.0 / aspect * tan(fovX / 2.0)));
+				double fovY = (2.0 * atan(1.0 / aspect * tan(fovX / 2.0)));
 				inlineCamera.setMember("fovY", FabricSplice::constructFloat64RTVal(fovY));
 			}
